feat: enabled username across profile, dropdown
